update home landing view
