+++ .dockerignore
@@ -15,6 +15,5 @@ packages/*
 !packages/www
 
 */*/dist
-*/*/docs
 */*/minio
 */*/out
