+++ .env.local.example
@@ -8,9 +8,6 @@ NEXT_PUBLIC_STRIPE_ENABLED_IN_DEV_MODE=
 SANITY_API_TOKEN=
 NEXT_PUBLIC_EMAIL_VERIFICATION_MODE=
 
-NEXT_PUBLIC_ALGOLIA_APP_ID=LOQ5Y01Z0L
-NEXT_PUBLIC_ALGOLIA_INDEX_NAME=docs
-NEXT_PUBLIC_ALGOLIA_SEARCH_API_KEY=2e5bcf93fb79be5661162b9509bf0a9f
 NEXT_PUBLIC_RECAPTCHA_SITE_KEY=
 NEXT_PUBLIC_HUBSPOT_STRIPE_FORM_ID=
 NEXT_PUBLIC_HUBSPOT_LOGIN_FORM_ID=
