+++ .vercelignore
@@ -1,4 +1,3 @@
 coverage*
-/docs
 .nyc_output
 /packages/api/data
