--- packages/www/components/Site/Login/index.tsx
@@ -1,198 +1,0 @@
-import { TextField, Grid, Box, Link as A } from "@livepeer/design-system";
-import { useEffect, useState } from "react";
-import hash from "@livepeer.studio/api/dist/hash";
-import { useRouter } from "next/router";
-import Button from "components/Site/Button";
-import { useHubspotForm } from "hooks";
-
-// The frontend salts are all the same. This could be configurable someday.
-export const FRONTEND_SALT = "69195A9476F08546";
-
-const Login = ({
-  id,
-  showName = false,
-  showOrganization = false,
-  showPhone = false,
-  showEmail,
-  showPassword,
-  buttonText,
-  onSubmit,
-  loading,
-  errors,
-}) => {
-  const router = useRouter();
-  const [firstName, setFirstName] = useState("");
-  const [lastName, setLastName] = useState("");
-  const [organization, setOrganization] = useState("");
-  const [phone, setPhone] = useState("");
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-
-  const { handleSubmit } = useHubspotForm({
-    portalId: process.env.NEXT_PUBLIC_HUBSPOT_PORTAL_ID,
-    formId: process.env.NEXT_PUBLIC_HUBSPOT_LOGIN_FORM_ID,
-  });
-
-  useEffect(() => {
-    if (router?.query?.email) {
-      setEmail(router.query.email as string);
-    }
-  }, [router?.query?.email]);
-
-  const submit = async (e) => {
-    e.preventDefault();
-    handleSubmit(e);
-
-    if (!showPassword) {
-      return onSubmit({ email });
-    }
-    const [hashedPassword] = await hash(password, FRONTEND_SALT);
-    // hash password, then
-    onSubmit({
-      email,
-      password: hashedPassword,
-      firstName,
-      lastName,
-      organization,
-      phone,
-    });
-  };
-
-  return (
-    <Box
-      css={{
-        position: "relative",
-        width: "100%",
-      }}>
-      <Box
-        as="form"
-        onSubmit={submit}
-        css={{
-          textAlign: "center",
-          width: "100%",
-          display: "flex",
-          flexDirection: "column",
-          alignItems: "center",
-          mb: "$3",
-          ml: "auto",
-          mr: "auto",
-          maxWidth: 500,
-        }}
-        id={id}>
-        {showName && (
-          <Grid
-            gap={3}
-            css={{
-              gridTemplateColumns: "1fr 1fr",
-              width: "100%",
-              alignItems: "center",
-            }}>
-            <TextField
-              size="3"
-              id="firstName"
-              css={{ width: "100%", mb: "$3" }}
-              name="firstName"
-              type="text"
-              placeholder="First name"
-              required
-              value={firstName}
-              onChange={(e) => setFirstName(e.target.value)}
-            />
-            <TextField
-              size="3"
-              id="lastName"
-              css={{ width: "100%", mb: "$3" }}
-              name="lastName"
-              type="text"
-              placeholder="Last name"
-              required
-              value={lastName}
-              onChange={(e) => setLastName(e.target.value)}
-            />
-          </Grid>
-        )}
-        {showOrganization && showPhone && (
-          <Grid
-            gap={3}
-            css={{
-              gridTemplateColumns: "1fr 1fr",
-              width: "100%",
-              alignItems: "center",
-            }}>
-            <TextField
-              size="3"
-              id="organization"
-              css={{ width: "100%", mb: "$3" }}
-              name="organization"
-              type="organization"
-              placeholder="Organization (optional)"
-              value={organization}
-              onChange={(e) => setOrganization(e.target.value)}
-            />
-            <TextField
-              size="3"
-              id="phone"
-              css={{ width: "100%", mb: "$3" }}
-              name="phone"
-              type="phone"
-              placeholder="Phone (optional)"
-              value={phone}
-              onChange={(e) => setPhone(e.target.value)}
-            />
-          </Grid>
-        )}
-        {showEmail && (
-          <TextField
-            size="3"
-            id="email"
-            css={{
-              width: "100%",
-              mb: "$3",
-              mx: "$2",
-              "@bp1": {
-                mx: "$4",
-              },
-            }}
-            name="email"
-            type="email"
-            placeholder="Email"
-            required
-            value={email}
-            onChange={(e) => setEmail(e.target.value)}
-          />
-        )}
-        {showPassword && (
-          <TextField
-            size="3"
-            id="password"
-            css={{
-              width: "100%",
-              mx: "$2",
-              "@bp1": {
-                mx: "$4",
-              },
-            }}
-            name="password"
-            type="password"
-            placeholder="Password"
-            required
-            value={password}
-            onChange={(e) => setPassword(e.target.value)}
-          />
-        )}
-
-        {errors.length > 0 && (
-          <Box css={{ mt: "$2" }}>{errors.join(", ")}&nbsp;</Box>
-        )}
-
-        <Button
-          small
-          css={{ width: "100%", my: "$3", px: "$3", fontSize: "$3" }}>
-          {loading ? "Loading..." : buttonText}
-        </Button>
-      </Box>
-    </Box>
-  );
-};
-
-export default Login;
