+++ packages/www/components/Site/Navigation/index.tsx
@@ -1,10 +1,5 @@
-import React, { useMemo, useState, useCallback, useEffect } from "react";
+import React from "react";
 import NavigationBase from "./base";
-import { useRouter } from "next/router";
-import DocsMobileSubMenu from "./mobile/docs-submenu";
-import { Tree } from "components/Site/TableOfContents";
-import { isMobile } from "react-device-detect";
-
 const defaultNavProps = {
   links: [
     {
@@ -36,82 +31,4 @@ const DefaultNav = ({ navBackgroundColor = "$loContrast" }) => (
   />
 );
 
-type DocsNavProps = {
-  tree?: Tree[];
-  ignoreList?: string[];
-};
-
-const DocsNav = ({ tree, ignoreList }: DocsNavProps) => {
-  const router = useRouter();
-  const { pathname } = router;
-  const [mobileSubmenuVisible, setMobileSubmenuVisible] = useState(true);
-  const [lastScrollTop, setLastScrollTop] = useState<number>();
-
-  const handleScroll = useCallback(() => {
-    const { scrollTop: currentScrollTop } = document.documentElement;
-    setLastScrollTop(currentScrollTop);
-    if (currentScrollTop > lastScrollTop) {
-      setMobileSubmenuVisible(false);
-    } else if (currentScrollTop < lastScrollTop) {
-      setMobileSubmenuVisible(true);
-    }
-  }, [lastScrollTop]);
-
-  useEffect(() => {
-    document.addEventListener("scroll", handleScroll);
-    return () => document.removeEventListener("scroll", handleScroll);
-  }, [handleScroll]);
-
-  const docsNavProps = useMemo(
-    () => ({
-      links: [
-        {
-          href: "/docs/guides",
-          children: "Docs",
-        },
-        {
-          href: "/docs/api-reference",
-          children: "API Reference",
-        },
-      ],
-      breadcrumb: [
-        {
-          children: "Docs",
-          href: "/docs/guides",
-          mobileDropdownLinks: [
-            {
-              children: "Docs",
-              href: "/docs/guides",
-              isSelected: pathname === "/docs/guides",
-            },
-            {
-              children: "API Reference",
-              href: "/docs/api-reference",
-              isSelected: pathname === "/docs/api-reference",
-            },
-          ],
-        },
-      ],
-    }),
-    [pathname]
-  );
-
-  return (
-    <>
-      <NavigationBase
-        {...docsNavProps}
-        withShadow={!mobileSubmenuVisible || !tree || !isMobile}
-        css={tree ? { backgroundColor: "$panel" } : undefined}
-      />
-      {tree && (
-        <DocsMobileSubMenu
-          tree={tree}
-          ignoreList={ignoreList}
-          mobileSubmenuVisible={true}
-        />
-      )}
-    </>
-  );
-};
-
-export { DefaultNav, DocsNav };
+export { DefaultNav };
