+++ packages/www/layouts/main.tsx
@@ -1,6 +1,4 @@
 import { Flex, Box } from "@livepeer/design-system";
-import { DefaultNav } from "components/Site/Navigation";
-import Footer from "components/Site/Footer";
 import ReactGA from "react-ga";
 import Router from "next/router";
 import { useEffect } from "react";
@@ -51,7 +49,6 @@ function Layout({
   noindex = false,
   preview = false,
   css = {},
-  navBackgroundColor = "transparent",
 }: Props) {
   useEffect(() => {
     if (window.location.hostname === "livepeer.studio") {
@@ -114,9 +111,7 @@ function Layout({
             Preview Mode
           </Box>
         )}
-        <DefaultNav navBackgroundColor={navBackgroundColor} />
         {children}
-        <Footer />
       </Flex>
     </>
   );
