--- packages/www/next-mdx.json
@@ -1,6 +1,0 @@
-{
-    "doc": {
-      "contentPath": "docs",
-      "basePath": "/"
-    }
-  }
