+++ packages/www/package.json
@@ -10,7 +10,6 @@
     "start": "next start -p 3012",
     "postbuild": "next-sitemap",
     "type-check": "tsc --pretty --noEmit",
-    "docs-positions": "node ./generate-docs-positions.js",
     "docker:build": "docker build $(printf ' -t livepeerci/www:%s' ${DOCKER_TAGS:-master}) -f Dockerfile ../..",
     "pkg": "pkg --out-path=bin --no-bytecode --public-packages \"*\" --public . --compress GZip ",
     "pkg:local": "yarn run build && yarn run pkg --targets=$(node -p '`node16-${os.platform()}-${os.arch()}`')",
@@ -18,12 +17,11 @@
   },
   "license": "MIT",
   "dependencies": {
-    "@docsearch/react": "3.0.0",
     "@emotion/core": "^10.0.17",
     "@emotion/react": "^11.4.0",
     "@emotion/styled": "^10.0.23",
     "@livepeer.studio/api": "^0.14.2",
-    "@livepeer/design-system": "^1.0.4",
+    "@livepeer/design-system": "^1.0.9",
     "@livepeer/react": "^2.4.0",
     "@mdx-js/loader": "^1.6.16",
     "@mdx-js/mdx": "^1.6.16",
