+++ packages/www/pages/_app.tsx
@@ -5,7 +5,6 @@ import { ApiProvider } from "hooks/use-api";
 import { AnalyzerProvider } from "hooks/use-analyzer";
 import { QueryClient, QueryClientProvider } from "react-query";
 import { MetaMaskProvider } from "metamask-react";
-import "../css/algolia-docsearch.css";
 import "../css/recaptcha.css";
 import React from "react";
 import {
@@ -31,13 +30,12 @@ const globalStyles = globalCss({
   body: {
     margin: 0,
     bc: "$loContrast",
-    fontFamily: "$untitled",
+    fontFamily: "Inter",
     color: "$hiContrast",
   },
 
   ".main": {
     bc: "$loContrast",
-    fontFamily: "Matter",
   },
 
   "h1, h2, h3, h4, h5": { fontWeight: 500 },
